requirements.txt
groq
python-dotenv
googlesearch-python